@@ -1,41 +1,41 @@
 {
-  "name": "fbe-on-web-react",
-  "version": "1.0.0",
-  "private": false,
-  "description": "フローチャートを作成・実行できるWebツールです。近日v2をリリース予定です",
-  "license": "MIT",
+  "name": "v2.0",
+  "version": "0.1.0",
+  "private": true,
   "dependencies": {
-    "-": "^0.0.1",
-    "@material-ui/core": "^4.12.3",
-    "@material-ui/icons": "^4.11.2",
-    "@mui/icons-material": "^5.0.4",
-    "@mui/material": "^5.0.4",
-    "@testing-library/jest-dom": "^5.11.4",
-    "@testing-library/react": "^11.1.0",
-    "@testing-library/user-event": "^12.1.10",
-    "@types/jest": "^26.0.15",
-    "@types/lodash": "^4.14.176",
-    "@types/node": "^12.0.0",
-    "@types/react": "^17.0.0",
-    "@types/react-dom": "^17.0.0",
-    "deta": "^1.0.1",
-    "faunadb": "^4.3.0",
-    "html-to-image": "^1.9.0",
+    "@emotion/react": "^11.7.1",
+    "@emotion/styled": "^11.6.0",
+    "@mui/icons-material": "^5.2.5",
+    "@mui/material": "^5.2.7",
+    "@mui/styles": "^5.2.3",
+    "@testing-library/jest-dom": "^5.16.1",
+    "@testing-library/react": "^12.1.2",
+    "@testing-library/user-event": "^13.5.0",
+    "@types/jest": "^27.4.0",
+    "@types/node": "^16.11.19",
+    "@types/react": "^17.0.38",
+    "@types/react-dom": "^17.0.11",
+    "@types/uuid": "^8.3.4",
+    "chevrotain": "^10.0.0",
     "lodash": "^4.17.21",
     "react": "^17.0.2",
+    "react-dnd": "^14.0.5",
+    "react-dnd-html5-backend": "^14.1.0",
     "react-dom": "^17.0.2",
-    "react-redux": "^7.2.4",
-    "react-scripts": "4.0.3",
-    "recoil": "^0.4.0",
-    "redux": "^4.1.1",
-    "store": "^2.0.12",
-    "styled-components": "^5.3.0",
-    "typescript": "^4.1.2",
-    "web-vitals": "^1.0.1"
+    "react-hotkeys-hook": "^3.4.4",
+    "react-redux": "^7.2.6",
+    "react-scripts": "5.0.0",
+    "redux": "^4.1.2",
+    "simple-eval": "^1.0.0",
+    "storejs": "^2.0.1",
+    "typescript": "^4.5.4",
+    "typescript-fsa": "^3.0.0",
+    "typescript-fsa-reducers": "^1.2.2",
+    "uuid": "^8.3.2",
+    "web-vitals": "^2.1.3"
   },
   "scripts": {
     "start": "react-scripts start",
-    "s": "react-scripts start",
     "build": "react-scripts build",
     "test": "react-scripts test",
     "eject": "react-scripts eject"
@@ -57,9 +57,5 @@
       "last 1 firefox version",
       "last 1 safari version"
     ]
-  },
-  "devDependencies": {
-    "@types/store": "^2.0.2",
-    "@types/styled-components": "^5.1.12"
   }
 }
