@@ -1,31 +1,22 @@
+import CssBaseline from '@mui/material/CssBaseline';
 import React from 'react';
 import ReactDOM from 'react-dom';
-
+import { Provider } from 'react-redux';
 import App from './App';
-// import reportWebVitals from './reportWebVitals';
+import { store } from './redux/store';
+import { DndProvider } from 'react-dnd'
+import { HTML5Backend } from 'react-dnd-html5-backend'
 
-import { Provider } from "react-redux"
-import {store} from "./redux/store";
-
-
-
+import "src/style/global.css" ;
 
 ReactDOM.render(
   <React.StrictMode>
-    <Provider store={store}>
-      <App/>
-    </Provider>
+    <DndProvider backend={HTML5Backend}>
+      <CssBaseline />
+      <Provider store={store}>
+        <App />
+      </Provider>
+    </DndProvider>
   </React.StrictMode>,
   document.getElementById('root')
 );
-
-// If you want to start measuring performance in your app, pass a function
-// to log results (for example: reportWebVitals(console.log))
-// or send to an analytics endpoint. Learn more: https://bit.ly/CRA-vitals
-
-// reportWebVitals((data)=>{
-//   console.log(
-//     "---------------",
-//     data,
-//     "---------------");
-// });
