@@ -1,67 +1,26 @@
-import { Item, Items } from "redux/types/item";
+import { ItemType } from "src/items/itemTypes";
+import actionCreatorFactory from "typescript-fsa";
+import { Item, ItemId, OptionValue } from "./types";
 
-export const actionTypes = {
-    add:"items/add",
-	set:"items/set",
-	remove:"items/remove",
-	option:{
-		set:"items/option/set",
-	},
-	sym:{
-		add:"items/sym/add",
-		remove:"items/sym/remove",
-	},
-	load:"items/load",
-	exchange:"items/exchange",
+const actionCreator = actionCreatorFactory();
 
-} ;
+export const setItem = actionCreator<{
+    item: Item;
+}>("items/set");
+export const removeItem = actionCreator<{
+    itemId: ItemId;
+}>("items/remove");
+export const setOption = actionCreator<{
+    itemId: ItemId;
+    name: string;
+    value: OptionValue;
+}>("items/options/set");
+export const setItemType = actionCreator<{
+    itemId:ItemId;
+    newItemType:ItemType;
+}>("items/itemType/set");
 
-//actionCreators
-export function addItem(item :Item) {
-    return {
-        type:actionTypes.add,
-        payload:item,
-    } ;
-}
-export function setItem(id :string, item :Item) {
-    return {
-        type:actionTypes.set,
-        payload:{
-            id,
-            item,
-        },
-    } ;
-}
-export function removeItem(id :string) {
-    return {
-        type:actionTypes.remove ,
-        payload:id ,
-    } ;
-}
-export function setOption(id :string, name :string, value: string|number|boolean){
-    return {
-        type:actionTypes.option.set ,
-        payload:{
-            id,
-            name,
-            value
-        },
-    } ;
-}
-export function loadItems(items :Items){
-    return {
-        type:actionTypes.load,
-        payload:items,
-    } ;
-}
-export function exchangeItem(itemId1:string, itemId2:string){
-    return {
-        type:actionTypes.exchange,
-        payload:{
-            itemId1,
-            itemId2
-        }
-    } ;
-}
-
-
+export const loadItems = actionCreator<{
+    items :Item[],
+}>("items/load");
+export const resetItems = actionCreator("items/reset")
