@@ -1,14 +1,11 @@
-
-import {
+import { 
+    combineReducers, 
     createStore, 
-    combineReducers,
+    compose, 
     applyMiddleware,
-    compose,
+    Middleware,
 } from "redux";
-
-import app from "./app/reducer" ;
-import items from "./items/reducer" ;
-import top from "./top/reducer" ;
+import { rootReducer } from "./root/reducer";
 
 interface ExtendedWindow extends Window {
     __REDUX_DEVTOOLS_EXTENSION_COMPOSE__?: typeof compose;
@@ -16,16 +13,16 @@ interface ExtendedWindow extends Window {
 declare var window: ExtendedWindow;
 const composeReduxDevToolsEnhancers = typeof window === 'object' && window.__REDUX_DEVTOOLS_EXTENSION_COMPOSE__ || compose;
 
-const reducer = combineReducers({
-    app,
-    items,
-    top,
-});
 
-// TODO: 適宜ミドルウェア定義
-const middlewares = [] as const;
 
-export const store = createStore( 
-    reducer,
-    composeReduxDevToolsEnhancers(applyMiddleware(...middlewares)),
-);
+const middlewares = [] as Middleware[] ;
+
+export const store = createStore(
+    rootReducer,
+    undefined,
+    composeReduxDevToolsEnhancers(applyMiddleware(...middlewares))
+) ;
+export type StoreState = ReturnType<typeof store.getState> ;
+
+
+
